requirements.txt
Django==1.7.0
SQLAlchemy==1.0.0
MySQL-python == 1.2.5
argparse==1.2.1
psycopg2==2.6
requests==2.6.0
wsgiref==0.1.2
djangorestframework>=2.4.3
django-oauth-plus>=2.2.1
oauth2>=1.5.211
django-oauth2-provider>=0.2.4